requirement.txt
streamlit
langchain
langchain-community
faiss-cpu
python-dotenv
unstructured
sentence-transformers
beautifulsoup4
requests
ollama
